Adding triangle example

+++ app/src/main/cpp/Camera.hpp
@@ -15,7 +15,7 @@ class Camera {
 
 public:
 
-    int w, h;
+    float w, h;
     float x, y, z;
     float ncp; // near clipping plane
     float fcp; // far clipping plane
